fix(MovieAPI): a static function is used to fetch images
