+++ BAZProject/Controllers/CategoryTableViewCell.swift
@@ -46,13 +46,14 @@ extension CategoryTableViewCell: UICollectionViewDelegate, UICollectionViewDataS
         let collectionCell = collectionView.dequeueReusableCell(withReuseIdentifier: "MovieGalleryCollectionViewCell", for: indexPath) as! MovieGalleryCollectionViewCell
         collectionCell.movieTitle.text = moviesToShow[indexPath.row].title
         collectionCell.voteAvarage.text = moviesToShow[indexPath.row].averageStars
-        
         let strImage =  moviesToShow[indexPath.row].poster_path
         let imageURL =  "https://image.tmdb.org/t/p/w500\(strImage)"
         if let url = URL(string: imageURL){
-            collectionCell.movieImage.fetchPhoto(url: url) { image, error in
+            MovieAPI.fetchPhoto(url: url) { image, error in
                 if let image = image{
                     collectionCell.movieImage.image = image
+                }else{
+                    collectionCell.movieImage.image = UIImage(named: "poster")
                 }
             }
         }
