+++ BAZProject/MovieAPI/MovieAPI.swift
@@ -5,6 +5,7 @@
 //
 
 import Foundation
+import UIKit
 
 class MovieAPI {
 
@@ -21,6 +22,36 @@ class MovieAPI {
             return []
         }
         return json.results
+    }
+    
+    /// Fetch movie poster with completion handler
+    ///
+    ///  - Parameter url: The given image URL
+    ///  - Returns: UIImage.
+    ///
+    
+    static func fetchPhoto(url: URL, completionHandler: @escaping (UIImage?, Error?) -> Void)
+    {
+        let task = URLSession.shared.dataTask(with: url) { (data, response, error) in
+            if let error = error{
+                DispatchQueue.main.async {
+                    completionHandler(nil, error)
+                }
+            }
+            if let data = data, let httpResponse = response as? HTTPURLResponse,
+               httpResponse.statusCode == 200 {
+                if let imageToShow = UIImage(data: data){
+                    DispatchQueue.main.async {
+                        completionHandler(imageToShow,nil)
+                    }
+                }else{
+                    DispatchQueue.main.async {
+                        completionHandler(UIImage(named: "poster"),nil)
+                    }
+                }
+            }
+        }
+        task.resume()
     }
     
 }
